Tests/LibWeb/Text/expected/cross-origin-isolated-with-coop-coep.txt

crossOriginIsolated: true
typeof SharedArrayBuffer: function
SharedArrayBuffer serialization threw: false
